+++ layouts/components/Cta.js
@@ -16,7 +16,7 @@ function Cta({ cta }) {
               height={206}
             />
           </div>
-          <div className="mt-5 text-center md:mt-0 md:text-left md:col-6 lg:col-5">
+          <div className="mt-5 text-center md:col-6 lg:col-5 md:mt-0 md:text-left">
             <h2>{cta?.title}</h2>
             <p className="mt-6">{markdownify(cta?.content)}</p>
             {cta.button.enable && (
