+++ layouts/components/YoutubePlayer.js
@@ -1,3 +1,5 @@
+"use client";
+
 import LiteYouTubeEmbed from "react-lite-youtube-embed";
 import "react-lite-youtube-embed/dist/LiteYouTubeEmbed.css";
