+++ layouts/partials/Header.js
@@ -1,13 +1,14 @@
+"use client";
+
 import Logo from "@components/Logo";
-import config from "@config/config.json";
 import menu from "@config/menu.json";
 import Link from "next/link";
-import { useRouter } from "next/router";
+import { usePathname } from "next/navigation";
 import React, { useState } from "react";
+import config from "../../config/config.json";
 
 const Header = () => {
-  //router
-  const router = useRouter();
+  const pathname = usePathname();
 
   // distructuring the main menu from menu object
   const { main } = menu;
@@ -30,7 +31,7 @@ const Header = () => {
         {/* navbar toggler */}
         <button
           id="show-button"
-          className="order-2 flex cursor-pointer items-center md:hidden md:order-1"
+          className="order-2 flex cursor-pointer items-center md:order-1 md:hidden"
           onClick={() => setNavOpen(!navOpen)}
         >
           {navOpen ? (
@@ -86,7 +87,7 @@ const Header = () => {
                       href={menu.url}
                       onClick={() => setNavOpen(false)}
                       className={`nav-link block ${
-                        router.asPath === menu.url ? "nav-link-active" : ""
+                        pathname === menu.url ? "nav-link-active" : ""
                       }`}
                     >
                       {menu.name}
@@ -109,7 +110,7 @@ const Header = () => {
           </ul>
         </div>
         {enable && (
-          <div className="d-flex order-1 ml-auto hidden min-w-[200px] items-center justify-end md:ml-0 md:flex md:order-2">
+          <div className="d-flex order-1 ml-auto hidden min-w-[200px] items-center justify-end md:order-2 md:ml-0 md:flex">
             <Link className="btn btn-primary z-0 py-[14px]" href={link} rel="">
               {label}
             </Link>
