--- pages/_document.js
@@ -1,40 +1,0 @@
-import config from "@config/config.json";
-import { Head, Html, Main, NextScript } from "next/document";
-
-const Document = () => {
-  // destructuring items from config object
-  const { favicon } = config.site;
-  return (
-    <Html lang="en">
-      <Head>
-        {/* favicon */}
-        <link
-          rel="shortcut icon"
-          type="image/png"
-          sizes="32x32"
-          href={favicon}
-        />
-        {/* theme meta */}
-        <meta name="theme-name" content="next-boilerplate" />
-        <meta name="msapplication-TileColor" content="#000000" />
-        <meta
-          name="theme-color"
-          media="(prefers-color-scheme: light)"
-          content="#fff"
-        />
-        <meta
-          name="theme-color"
-          media="(prefers-color-scheme: dark)"
-          content="#000"
-        />
-      </Head>
-      <body>
-        <Main />
-        {/* <TwSizeIndicator /> */}
-        <NextScript />
-      </body>
-    </Html>
-  );
-};
-
-export default Document;
